Refactor numerical computations to use double precision for checksum calculations and array initializations across multiple files, enhancing accuracy in simulations.

+++ llama2-transformer/transformer.c
@@ -424,7 +424,7 @@ int main(void) {
   printf("\n");
 
   // Verification
-  float checksum = 0.0f;
+  double checksum = 0.0;
   for (int i = 0; i < VOCAB_SIZE; i++) {
     checksum += logits_out[i];
   }
